feat: split main wizzard into actual web pages
