+++ packages_rs/nextclade-web/src/pages/dataset-suggest.tsx
@@ -1,0 +1,1 @@
+export { DatasetAutodetectPage as default } from 'src/components/Main/DatasetAutodetectPage'
