+++ packages_rs/nextclade-web/src/pages/dataset.tsx
@@ -1,0 +1,1 @@
+export { DatasetPage as default } from 'src/components/Main/DatasetPage'
